@@ -3,7 +3,7 @@ description: A new Flutter project.
 publish_to: 'none'
 version: 1.0.0+1
 environment:
-  sdk: ">=2.9.0 <3.0.0"
+  sdk: ">=2.10.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -28,6 +28,9 @@ dependencies:
   geocoding: ^1.0.5
   google_maps_flutter: ^1.2.0
   uuid: ^2.2.2
+  cached_network_image: ^2.5.0
+  timeago: ^2.0.29
+  share: ^0.6.5+4
 
 
 dev_dependencies:
