@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:min3_twitwi/common/style.dart';
+import 'package:min3_twitwi/enum/constant.dart';
 import 'package:min3_twitwi/generated/l10n.dart';
+import 'package:min3_twitwi/view/common/style.dart';
+import 'package:min3_twitwi/viewmodel/feed_view_model.dart';
 import 'package:min3_twitwi/viewmodel/post_view_model.dart';
 import 'package:provider/provider.dart';
 
@@ -8,7 +10,10 @@ import 'package:provider/provider.dart';
 
 
 class PostCaptionInputTextField extends StatefulWidget {
-  PostCaptionInputTextField({Key key}) : super(key: key);
+  final String captionBeforeUpload;   /// [caption編集画面への遷移時に、現行caption見たい]
+  final PostCaptionOpenMode postCaptionOpenMode;   /// [Feed/Postどっちから開くか]
+  PostCaptionInputTextField({this.captionBeforeUpload, this.postCaptionOpenMode});
+
   @override
   _PostCaptionInputTextFieldState createState() => _PostCaptionInputTextFieldState();
 }
@@ -24,6 +29,10 @@ class _PostCaptionInputTextFieldState extends State<PostCaptionInputTextField> {
     _captionTextEditingController.addListener(() {   /// [(() { }): VoidCallback]
       _onCaptionUpdated();
     });
+
+    if (widget.postCaptionOpenMode == PostCaptionOpenMode.FROM_FEED) {
+      _captionTextEditingController.text = widget.captionBeforeUpload;
+    }
   }
   @override
   void dispose() {
@@ -49,10 +58,15 @@ class _PostCaptionInputTextFieldState extends State<PostCaptionInputTextField> {
 
 
   _onCaptionUpdated() {
-    final postViewModel = Provider.of<PostViewModel>(context, listen: false);
-    /// [input内容 -> viewModelに格納]
-    postViewModel.caption = _captionTextEditingController.text;
-    // print("comm120: _onCaptionUpdated(): ${postViewModel.caption}");
-    print("comm120: _onCaptionUpdated()");
+    if (widget.postCaptionOpenMode == PostCaptionOpenMode.FROM_FEED) {
+        final feedViewModel = Provider.of<FeedViewModel>(context, listen: false);
+        feedViewModel.caption = _captionTextEditingController.text;
+        print("comm120: _onCaptionUpdated(): PostCaptionOpenMode.FROM_FEED");
+    } else {
+        final postViewModel = Provider.of<PostViewModel>(context, listen: false);
+        /// [input内容 -> viewModelに格納]
+        postViewModel.caption = _captionTextEditingController.text;
+        print("comm121: _onCaptionUpdated(): PostCaptionOpenMode.FROM_POST");
+    }
   }
 }
